reutilizamos codigo y minimo 4 caracteres
